# internal/models/comms/types.gen.go
// Package comms provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.1-0.20260518235555-9bb79e703422 DO NOT EDIT.
package comms

import (
	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerAuthScopes bearerAuthContextKey = "bearerAuth.Scopes"
)

// Defines values for SendResultStatus.
const (
	Failed SendResultStatus = "failed"
	Queued SendResultStatus = "queued"
	Sent   SendResultStatus = "sent"
)

// Valid indicates whether the value is a known member of the SendResultStatus enum.
func (e SendResultStatus) Valid() bool {
	switch e {
	case Failed:
		return true
	case Queued:
		return true
	case Sent:
		return true
	default:
		return false
	}
}

// Problem RFC 7807 Problem+JSON
type Problem struct {
	Detail   *string `json:"detail,omitempty"`
	Instance *string `json:"instance,omitempty"`
	Status   int     `json:"status"`
	Title    string  `json:"title"`
	Type     string  `json:"type"`
}

// SendEmailRequest defines model for SendEmailRequest.
type SendEmailRequest struct {
	// Locale BCP-47 locale tag (e.g. en-US). Defaults to en-US.
	Locale  *string              `json:"locale,omitempty"`
	ReplyTo *openapi_types.Email `json:"reply_to,omitempty"`

	// TemplateData Key/value pairs injected into the template at render time
	TemplateData *map[string]interface{} `json:"template_data,omitempty"`

	// TemplateId Template identifier (resolved from S3 template store)
	TemplateId string `json:"template_id"`

	// To Recipient email address
	To openapi_types.Email `json:"to"`
}

// SendPushRequest defines model for SendPushRequest.
type SendPushRequest struct {
	Body string `json:"body"`

	// Data Arbitrary key/value payload delivered alongside the notification
	Data *map[string]interface{} `json:"data,omitempty"`

	// DeviceToken FCM or APNs device token
	DeviceToken string `json:"device_token"`
	Title       string `json:"title"`
}

// SendResult defines model for SendResult.
type SendResult struct {
	// MessageId Provider-assigned message identifier for tracking
	MessageId string           `json:"message_id"`
	Status    SendResultStatus `json:"status"`
}

// SendResultStatus defines model for SendResult.Status.
type SendResultStatus string

// SendSmsRequest defines model for SendSmsRequest.
type SendSmsRequest struct {
	Body string `json:"body"`

	// To E.164 formatted phone number (e.g. +12125551234)
	To string `json:"to"`
}

// BadRequest RFC 7807 Problem+JSON
type BadRequest = Problem

// Unauthorized RFC 7807 Problem+JSON
type Unauthorized = Problem

// bearerAuthContextKey is the context key for bearerAuth security scheme
type bearerAuthContextKey string

// SendEmailJSONRequestBody defines body for SendEmail for application/json ContentType.
type SendEmailJSONRequestBody = SendEmailRequest

// SendPushJSONRequestBody defines body for SendPush for application/json ContentType.
type SendPushJSONRequestBody = SendPushRequest

// SendSmsJSONRequestBody defines body for SendSms for application/json ContentType.
type SendSmsJSONRequestBody = SendSmsRequest