# openapi.yaml
openapi: "3.1.2"
info:
  title: komodo-auth-api
  version: "0.1.0"
  description: |
    OAuth 2.0 authorization server. Issues and validates RS256 JWTs.

    **Two servers:**
    - Public (7011): client-facing OAuth endpoints + JWKS
    - Internal (7012): token validation and client registry for service-to-service calls

    **HTTP conventions:**
    - POST → 201 Created / 400 Bad Request / 409 Conflict
    - PUT → 200 OK / 404 Not Found
    - DELETE → 204 No Content / 404 Not Found
    - GET → 200 OK / 404 Not Found
servers:
  - url: http://localhost:7011
    description: Public (local)
  - url: http://localhost:7012
    description: Internal (local)
...
paths:
...
  /v1/oauth/authorize:
    get:
      tags: [OAuth, Public]
      summary: Authorization endpoint (authorization_code flow)
      operationId: authorize
      security: []
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum: [code]
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: scope
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: code_challenge
          in: query
          schema:
            type: string
            description: PKCE code challenge — BASE64URL(SHA256(code_verifier)); required for public/browser clients (RFC 7636)
        - name: code_challenge_method
          in: query
          schema:
            type: string
            enum: [S256]
            description: PKCE challenge method — only S256 is accepted
      responses:
        "302":
          description: Redirect to client with authorization code
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /v1/otp/verify:
    post:
      tags: [OTP, Public]
      summary: Verify a one-time password and issue a short-lived token
      operationId: verifyOTP
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OTPVerifyRequest"
      responses:
        "200":
          description: OTP valid — returns a short-lived JWT or session token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    ClientJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Service/client JWT — required for introspect

    UserJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User access token (OTP-verified or passkey-issued) — required for passkey registration
  schemas:
...
    TokenResponse:
      type: object
      # RFC 6749 §5.1 token response (snake_case).
      required: [access_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: Seconds until token expires
        refresh_token:
          type: string
        scope:
          type: string
...
    OTPVerifyRequest:
      type: object
      required: [email, code]
      properties:
        email:
          type: string
          format: email
        code:
          type: string
          description: The OTP code delivered to the user
...
    Problem:
      type: object
      description: RFC 7807 Problem+JSON
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...